Code refactoring and optimiztations
