--- src/Components/RequiredItem.js
@@ -1,9 +1,0 @@
-const RequiredItem = ({img_src}) => {
-    return (
-        <div className="card">
-            <img src={img_src} alt="required item" />
-        </div>
-    )
-}
-
-export default RequiredItem
