+++ src/Components/RequiredItems.js
@@ -1,4 +1,3 @@
-import RequiredItem from './RequiredItem'
 import Cap from '../assets/img/cap.png'
 import Gloves from '../assets/img/gloves.png' 
 import Jacket from '../assets/img/jacket.png'
@@ -47,9 +46,17 @@ const RequiredItems = ({ weatherKind }) => {
     }
 
     return (
-        <div className='items-card'>
-            {getItems(weatherKind) !== undefined && getItems(weatherKind).map((ele, index) => { 
-                return <RequiredItem img_src={ele} key={index} />})}
+        <div className='items-card-container'>
+            {
+                getItems(weatherKind) !== undefined && 
+                getItems(weatherKind).map((ele, index) => { 
+                    return (
+                        <div key={index} className="items-card">
+                            <img src={ele} alt="required item" />
+                        </div>
+                    )
+                })
+            }
         </div>
     )
 }
